requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
accelerate==0.23.0
    # via -r requirements.in
aiohttp==3.8.5
    # via
    #   cohere
    #   langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
annoy==1.17.3
    # via -r requirements.in
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
backoff==2.2.1
    # via cohere
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
cohere==4.24
    # via -r requirements.in
dataclasses-json==0.5.14
    # via langchain
fastavro==1.8.2
    # via cohere
filelock==3.12.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.9.1
    # via huggingface-hub
greenlet==2.0.2
    # via sqlalchemy
huggingface-hub==0.17.2
    # via
    #   accelerate
    #   transformers
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via cohere
jinja2==3.1.2
    # via torch
langchain==0.0.285
    # via -r requirements.in
langsmith==0.0.38
    # via langchain
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via torch
numexpr==2.8.6
    # via langchain
numpy==1.23.3
    # via
    #   -r requirements.in
    #   accelerate
    #   langchain
    #   numexpr
    #   pandas
    #   transformers
packaging==23.1
    # via
    #   accelerate
    #   huggingface-hub
    #   marshmallow
    #   transformers
pandas==2.1.0
    # via -r requirements.in
psutil==5.9.5
    # via accelerate
pydantic==2.3.0
    # via
    #   langchain
    #   langsmith
pydantic-core==2.6.3
    # via pydantic
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   accelerate
    #   huggingface-hub
    #   langchain
    #   transformers
regex==2023.8.8
    # via transformers
requests==2.31.0
    # via
    #   cohere
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   transformers
safetensors==0.3.3
    # via transformers
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.21
    # via langchain
sympy==1.12
    # via torch
tenacity==8.2.3
    # via langchain
tokenizers==0.13.3
    # via transformers
torch==2.0.1
    # via
    #   -r requirements.in
    #   accelerate
tqdm==4.66.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.33.2
    # via -r requirements.in
typing-extensions==4.8.0
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
urllib3==2.0.5
    # via
    #   cohere
    #   requests
yarl==1.9.2
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata
